pyproject.toml
[project]
name = "phantom-flow"
version = "0.1.0"
description = "Evidence-backed public funding zombie recipient triage dashboard."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "anthropic>=0.49.0",
  "httpx>=0.27.0",
  "pandas>=2.2.0",
  "python-dotenv>=1.0.1",
  "rapidfuzz>=3.9.0",
  "streamlit>=1.36.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]

[project.scripts]
phantom-flow = "phantom_flow.pipeline:main"

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
norecursedirs = ["pytest-cache-files-*", ".pytest_cache", ".git", ".venv"]
